+++ qoqo-qryd/src/lib.rs
@@ -101,7 +101,7 @@ fn qoqo_qryd(_py: Python, module: &PyModule) -> PyResult<()> {
     module.add_wrapped(wrapper)?;
     let wrapper = wrap_pymodule!(api_devices::api_devices);
     module.add_wrapped(wrapper)?;
-    let wrapper = wrap_pymodule!(tweezer_devices::experimental_devices);
+    let wrapper = wrap_pymodule!(tweezer_devices::tweezer_devices);
     module.add_wrapped(wrapper)?;
     // Adding nice imports corresponding to maturin example
     let wrapper = wrap_pymodule!(pragma_operations::pragma_operations);
@@ -116,8 +116,8 @@ fn qoqo_qryd(_py: Python, module: &PyModule) -> PyResult<()> {
     system_modules.set_item("qoqo_qryd.qryd_devices", module.getattr("qryd_devices")?)?;
     system_modules.set_item("qoqo_qryd.api_devices", module.getattr("api_devices")?)?;
     system_modules.set_item(
-        "qoqo_qryd.experimental_devices",
-        module.getattr("experimental_devices")?,
+        "qoqo_qryd.tweezer_devices",
+        module.getattr("tweezer_devices")?,
     )?;
     Ok(())
 }