+++ qoqo-qryd/src/pragma_operations.rs
@@ -24,7 +24,7 @@ use qoqo::operations::PragmaChangeDeviceWrapper;
 use roqoqo::prelude::*;
 use roqoqo_qryd::{
     PragmaChangeQRydLayout, PragmaDeactivateQRydQubit, PragmaShiftQRydQubit,
-    PragmaShiftQubitsTweezers,
+    PragmaShiftQubitsTweezers, PragmaSwitchDeviceLayout,
 };
 use std::collections::HashMap;
 
@@ -43,7 +43,7 @@ use std::collections::HashMap;
 /// Args:
 ///     new_layout (int): The index of the new layout.
 pub struct PragmaChangeQRydLayoutWrapper {
-    /// PragmaStateVector to be wrapped and converted to Python.
+    /// PragmaChangeQRydLayout to be wrapped and converted to Python.
     pub internal: PragmaChangeQRydLayout,
 }
 
@@ -142,8 +142,9 @@ impl PragmaChangeQRydLayoutWrapper {
             .map_err(|_| PyTypeError::new_err("Input cannot be converted to byte array"))?;
 
         Ok(PragmaChangeQRydLayoutWrapper {
-            internal: deserialize(&bytes[..])
-                .map_err(|_| PyValueError::new_err("Input cannot be deserialized to Circuit"))?,
+            internal: deserialize(&bytes[..]).map_err(|_| {
+                PyValueError::new_err("Input cannot be deserialized to PragmaChangeQRydLayout")
+            })?,
         })
     }
 
@@ -1049,6 +1050,263 @@ impl PragmaShiftQubitsTweezersWrapper {
     }
 }
 
+#[pyclass(
+    name = "PragmaSwitchDeviceLayout",
+    module = "qoqo_qryd.pragma_operations"
+)]
+#[derive(Clone, Debug, PartialEq, Eq)]
+/// This PRAGMA operation changes the layout of a Tweezer device.
+///
+/// Before running a circuit a number of layouts can be registered
+/// in the device with the `add_layout` method.
+///
+/// This PRAGMA operation switches between the predefined operations.
+///
+/// Args:
+///     new_layout (str): The name of the new layout.
+pub struct PragmaSwitchDeviceLayoutWrapper {
+    /// PragmaSwitchDeviceLayout to be wrapped and converted to Python.
+    pub internal: PragmaSwitchDeviceLayout,
+}
+
+#[pymethods]
+impl PragmaSwitchDeviceLayoutWrapper {
+    /// Create a PragmaSwitchDeviceLayout.
+    ///
+    /// Args:
+    ///     new_layout (str): The new layout the device is changed to.
+    ///
+    /// Returns:
+    ///     self: The new PragmaSwitchDeviceLayout.
+    #[new]
+    #[pyo3(text_signature = "(new_layout, /)")]
+    fn new(new_layout: String) -> Self {
+        Self {
+            internal: PragmaSwitchDeviceLayout::new(new_layout),
+        }
+    }
+
+    /// Return the name of the new layout the Pragma changes the device to.
+    ///
+    /// Returns:
+    ///     int: The name of the layout.
+    fn new_layout(&self) -> String {
+        self.internal.new_layout().clone()
+    }
+
+    /// Wrap PragmaSwitchDeviceLayout in PragmaChangeDevice operation
+    ///
+    /// PragmaSwitchDeviceLayout is device specific and can not be directly added to a Circuit.
+    /// Instead it is first wrapped in a PragmaChangeDevice operation that is in turn added
+    /// to the circuit.
+    ///
+    /// Example
+    /// -------
+    ///
+    /// >>> from qoqo import Circuit
+    /// ... from qoqo_qryd.pragma_operations import PragmaSwitchDeviceLayout
+    /// ... circuit = Circuit()
+    /// ... circuit += PragmaSwitchDeviceLayout(new_layout="Square").to_pragma_change_device()
+    ///
+    /// Returns:
+    ///     PragmaChangeDevice
+    pub fn to_pragma_change_device(&self) -> PyResult<PragmaChangeDeviceWrapper> {
+        Ok(PragmaChangeDeviceWrapper {
+            internal: self.internal.to_pragma_change_device().map_err(|err| {
+                PyRuntimeError::new_err(format!(
+                    "Error occured during serialisation of PragmaSwitchDeviceLayout {:?}",
+                    err
+                ))
+            })?,
+        })
+    }
+
+    /// List all involved qubits (here, all).
+    ///
+    /// Returns:
+    ///     set[int]: The involved qubits of the PRAGMA operation.
+    fn involved_qubits(&self) -> PyObject {
+        let pyobject: PyObject =
+            Python::with_gil(|py| -> PyObject { PySet::new(py, &["All"]).unwrap().to_object(py) });
+        pyobject
+    }
+
+    /// Return the bincode representation of the PragmaSwitchDeviceLayout using the bincode crate.
+    ///
+    /// Returns:
+    ///     ByteArray: The serialized Circuit (in bincode form).
+    ///
+    /// Raises:
+    ///     ValueError: Cannot serialize PragmaSwitchDeviceLayout to bytes.
+    pub fn to_bincode(&self) -> PyResult<Py<PyByteArray>> {
+        let serialized = serialize(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize Circuit to bytes"))?;
+        let b: Py<PyByteArray> = Python::with_gil(|py| -> Py<PyByteArray> {
+            PyByteArray::new(py, &serialized[..]).into()
+        });
+        Ok(b)
+    }
+
+    /// Convert the bincode representation of the PragmaSwitchDeviceLayout to
+    /// a PragmaSwitchDeviceLayout using the bincode crate.
+    ///
+    /// Args:
+    ///     input (ByteArray): The serialized PragmaSwitchDeviceLayout (in bincode form).
+    ///
+    /// Returns:
+    ///     PragmaSwitchDeviceLayout: The deserialized PragmaSwitchDeviceLayout.
+    ///
+    /// Raises:
+    ///     TypeError: Input cannot be converted to byte array.
+    ///     ValueError: Input cannot be deserialized to PragmaSwitchDeviceLayout.
+    pub fn from_bincode(&self, input: &PyAny) -> PyResult<PragmaSwitchDeviceLayoutWrapper> {
+        let bytes = input
+            .extract::<Vec<u8>>()
+            .map_err(|_| PyTypeError::new_err("Input cannot be converted to byte array"))?;
+
+        Ok(PragmaSwitchDeviceLayoutWrapper {
+            internal: deserialize(&bytes[..]).map_err(|_| {
+                PyValueError::new_err("Input cannot be deserialized to PragmaSwitchDeviceLayout")
+            })?,
+        })
+    }
+
+    /// Return tags classifying the type of the operation.
+    ///
+    /// Used for the type based dispatch in ffi interfaces.
+    ///
+    /// Returns:
+    ///     list[str]: The tags of the operation.
+    fn tags(&self) -> Vec<String> {
+        self.internal.tags().iter().map(|s| s.to_string()).collect()
+    }
+
+    /// Return hqslang name of the operation.
+    ///
+    /// Returns:
+    ///     str: The hqslang name of the operation.
+    fn hqslang(&self) -> &'static str {
+        self.internal.hqslang()
+    }
+
+    /// Return true when the operation has symbolic parameters.
+    ///
+    /// Returns:
+    ///     bool: True if the operation contains symbolic parameters, False if it does not.
+    fn is_parametrized(&self) -> bool {
+        self.internal.is_parametrized()
+    }
+
+    /// Substitute the symbolic parameters in a clone of the PRAGMA operation according to the substitution_parameters input.
+    ///
+    /// Args:
+    ///     substitution_parameters (dict[str, float]): The dictionary containing the substitutions to use in the PRAGMA operation.
+    ///
+    /// Returns:
+    ///     self: The PRAGMA operation operation with the parameters substituted.
+    ///
+    /// Raises:
+    ///     RuntimeError: The parameter substitution failed.
+    #[pyo3(text_signature = "(substitution_parameters, /)")]
+    fn substitute_parameters(
+        &self,
+        substitution_parameters: std::collections::HashMap<&str, f64>,
+    ) -> PyResult<Self> {
+        let mut calculator = qoqo_calculator::Calculator::new();
+        for (key, val) in substitution_parameters.iter() {
+            calculator.set_variable(key, *val);
+        }
+        Ok(Self {
+            internal: self
+                .internal
+                .substitute_parameters(&calculator)
+                .map_err(|x| {
+                    pyo3::exceptions::PyRuntimeError::new_err(format!(
+                        "Parameter Substitution failed: {:?}",
+                        x
+                    ))
+                })?,
+        })
+    }
+
+    /// Remap qubits in a clone of the PRAGMA operation.
+    ///
+    /// Args:
+    ///     mapping (dict[int, int]): The dictionary containing the {qubit: qubit} mapping to use in the PRAGMA operation.
+    ///
+    /// Returns:
+    ///     self: The PRAGMA operation with the qubits remapped.
+    ///
+    /// Raises:
+    ///     RuntimeError: The qubit remapping failed.
+    #[pyo3(text_signature = "(mapping, /)")]
+    fn remap_qubits(&self, mapping: std::collections::HashMap<usize, usize>) -> PyResult<Self> {
+        let new_internal = self
+            .internal
+            .remap_qubits(&mapping)
+            .map_err(|_| pyo3::exceptions::PyRuntimeError::new_err("Qubit remapping failed: "))?;
+        Ok(Self {
+            internal: new_internal,
+        })
+    }
+
+    /// Return a copy of the PRAGMA operation (copy here produces a deepcopy).
+    ///
+    /// Returns:
+    ///     PragmaChangeQRydLayout: A deep copy of self.
+    fn __copy__(&self) -> PragmaSwitchDeviceLayoutWrapper {
+        self.clone()
+    }
+
+    /// Return a deep copy of the PRAGMA operation.
+    ///
+    /// Returns:
+    ///     PragmaChangeQRydLayout: A deep copy of self.
+    fn __deepcopy__(&self, _memodict: Py<PyAny>) -> PragmaSwitchDeviceLayoutWrapper {
+        self.clone()
+    }
+
+    /// Return a string containing a formatted (string) representation of the PRAGMA operation.
+    ///
+    /// Returns:
+    ///     str: The string representation of the operation.
+    fn __format__(&self, _format_spec: &str) -> PyResult<String> {
+        Ok(format!("{:?}", self.internal))
+    }
+
+    /// Return a string containing a printable representation of the PRAGMA operation.
+    ///
+    /// Returns:
+    ///     str: The printable string representation of the operation.
+    fn __repr__(&self) -> PyResult<String> {
+        Ok(format!("{:?}", self.internal))
+    }
+
+    /// Return the __richcmp__ magic method to perform rich comparison operations on PragmaSwitchDeviceLayout.
+    ///
+    /// Args:
+    ///     self: The PragmaSwitchDeviceLayout object.
+    ///     other: The object to compare self to.
+    ///     op: Whether they should be equal or not.
+    ///
+    /// Returns:
+    ///     bool: Whether the two operations compared evaluated to True or False.
+    fn __richcmp__(&self, other: Py<PyAny>, op: pyo3::class::basic::CompareOp) -> PyResult<bool> {
+        let other: PragmaSwitchDeviceLayoutWrapper =
+            Python::with_gil(|py| -> PyResult<PragmaSwitchDeviceLayoutWrapper> {
+                let other_extracted: PyResult<PragmaSwitchDeviceLayoutWrapper> = other.extract(py);
+                other_extracted
+            })?;
+        match op {
+            pyo3::class::basic::CompareOp::Eq => Ok(self.internal == other.internal),
+            pyo3::class::basic::CompareOp::Ne => Ok(self.internal != other.internal),
+            _ => Err(pyo3::exceptions::PyNotImplementedError::new_err(
+                "Other comparison not implemented.",
+            )),
+        }
+    }
+}
+
 /// QRyd specific PragmaOperations that support changing the QRyd device during a circuit evaluation
 ///
 /// .. autosummary::
@@ -1058,11 +1316,13 @@ impl PragmaShiftQubitsTweezersWrapper {
 ///    PragmaShiftQRydQubit
 ///    PragmaDeactivateQRydQubit
 ///    PragmaShiftQubitsTweezers
+///    PragmaSwitchDeviceLayout
 #[pymodule]
 pub fn pragma_operations(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<PragmaChangeQRydLayoutWrapper>()?;
     m.add_class::<PragmaShiftQRydQubitWrapper>()?;
     m.add_class::<PragmaDeactivateQRydQubitWrapper>()?;
     m.add_class::<PragmaShiftQubitsTweezersWrapper>()?;
+    m.add_class::<PragmaSwitchDeviceLayoutWrapper>()?;
     Ok(())
 }