+++ roqoqo-qryd/src/tweezer_devices.rs
@@ -26,7 +26,10 @@ use roqoqo::{
     RoqoqoBackendError,
 };
 
-use crate::{phi_theta_relation, PragmaChangeQRydLayout, PragmaDeactivateQRydQubit};
+use crate::{
+    phi_theta_relation, PragmaDeactivateQRydQubit, PragmaShiftQubitsTweezers,
+    PragmaSwitchDeviceLayout,
+};
 
 /// Tweezer Device
 ///
@@ -58,6 +61,13 @@ pub struct TweezerLayoutInfo {
     pub tweezer_three_qubit_gate_times: HashMap<String, HashMap<(usize, usize, usize), f64>>,
     /// Maps a multi-qubit gate name to a Vec<tweezer> -> time mapping
     pub tweezer_multi_qubit_gate_times: HashMap<String, HashMap<Vec<usize>, f64>>,
+    /// Allowed shifts from one tweezer to others.
+    /// The keys give the tweezer a qubit can be shifted out of.
+    /// The values are lists over the directions the qubit in the tweezer can be shifted into.
+    /// The items in the list give the allowed tweezers the qubit can be shifted into in order.
+    /// For a list 1,2,3 the qubit can be shifted into tweezer 1, into tweezer 2 if tweezer 1 is not occupied,
+    /// and into tweezer 3 if tweezer 1 and 2 are not occupied.
+    pub allowed_tweezer_shifts: HashMap<usize, Vec<Vec<usize>>>,
 }
 
 #[derive(Debug, Clone, PartialEq, serde::Deserialize, serde::Serialize)]
@@ -70,6 +80,8 @@ struct TweezerLayoutInfoSerialize {
     tweezer_three_qubit_gate_times: Vec<(String, ThreeTweezersTimes)>,
     /// Maps a multi-qubit gate name to a Vec<tweezer> -> time mapping
     tweezer_multi_qubit_gate_times: Vec<(String, MultiTweezersTimes)>,
+    /// Allowed shifts from one tweezer to others
+    allowed_tweezer_shifts: Vec<(usize, Vec<Vec<usize>>)>,
 }
 type SingleTweezerTimes = Vec<(usize, f64)>;
 type TwoTweezersTimes = Vec<((usize, usize), f64)>;
@@ -98,12 +110,15 @@ impl From<TweezerLayoutInfoSerialize> for TweezerLayoutInfo {
             .into_iter()
             .map(|(k, v)| (k, v.into_iter().map(|(i_k, i_v)| (i_k, i_v)).collect()))
             .collect();
+        let allowed_tweezer_shifts: HashMap<usize, Vec<Vec<usize>>> =
+            info.allowed_tweezer_shifts.into_iter().collect();
 
         Self {
             tweezer_single_qubit_gate_times,
             tweezer_two_qubit_gate_times,
             tweezer_three_qubit_gate_times,
             tweezer_multi_qubit_gate_times,
+            allowed_tweezer_shifts,
         }
     }
 }
@@ -130,11 +145,15 @@ impl From<TweezerLayoutInfo> for TweezerLayoutInfoSerialize {
             .into_iter()
             .map(|(k, v)| (k, v.into_iter().map(|(i_k, i_v)| (i_k, i_v)).collect()))
             .collect();
+        let allowed_tweezer_shifts: Vec<(usize, Vec<Vec<usize>>)> =
+            info.allowed_tweezer_shifts.into_iter().collect();
+
         Self {
             tweezer_single_qubit_gate_times,
             tweezer_two_qubit_gate_times,
             tweezer_three_qubit_gate_times,
             tweezer_multi_qubit_gate_times,
+            allowed_tweezer_shifts,
         }
     }
 }
@@ -337,7 +356,7 @@ impl TweezerDevice {
         qubit: usize,
         tweezer: usize,
     ) -> Result<HashMap<usize, usize>, RoqoqoBackendError> {
-        if !self.is_tweezer_present(tweezer) {
+        if !self.is_tweezer_present(tweezer, None) {
             return Err(RoqoqoBackendError::GenericError {
                 msg: "The given tweezer is not present in the device Tweezer data.".to_string(),
             });
@@ -483,6 +502,57 @@ impl TweezerDevice {
         }
     }
 
+    /// Set the allowed Tweezer shifts of a specified Tweezer.
+    ///
+    /// The tweezer give the tweezer a qubit can be shifted out of. The values are lists
+    /// over the directions the qubit in the tweezer can be shifted into.
+    /// The items in the list give the allowed tweezers the qubit can be shifted into in order.
+    /// For a list 1,2,3 the qubit can be shifted into tweezer 1, into tweezer 2 if tweezer 1 is not occupied,
+    /// and into tweezer 3 if tweezer 1 and 2 are not occupied.
+    ///
+    /// # Arguments
+    ///
+    /// * `tweezer` - The index of the tweezer.
+    /// * `allowed_shifts` - The allowed Tweezer shifts.
+    /// * `layout_name` - The name of the Layout to apply the gate time in. Defaults to the current Layout.
+    pub fn set_allowed_tweezer_shifts(
+        &mut self,
+        tweezer: &usize,
+        allowed_shifts: &[&[usize]],
+        layout_name: Option<String>,
+    ) -> Result<(), RoqoqoBackendError> {
+        let layout_name = layout_name
+            .clone()
+            .unwrap_or_else(|| self.current_layout.clone());
+
+        if !self.is_tweezer_present(*tweezer, Some(layout_name.clone()))
+            || allowed_shifts.iter().any(|s| {
+                s.iter()
+                    .any(|t| !self.is_tweezer_present(*t, Some(layout_name.clone())))
+            })
+        {
+            return Err(RoqoqoBackendError::GenericError {
+                msg: "The given tweezer, or shifts tweezers, are not present in the device Tweezer data."
+                    .to_string(),
+            });
+        }
+        if allowed_shifts
+            .iter()
+            .any(|shift_list| shift_list.contains(tweezer))
+        {
+            return Err(RoqoqoBackendError::GenericError {
+                msg: "The allowed shifts contain the given tweezer.".to_string(),
+            });
+        }
+        if let Some(info) = self.layout_register.get_mut(&layout_name) {
+            info.allowed_tweezer_shifts.insert(
+                *tweezer,
+                allowed_shifts.iter().map(|&slice| slice.to_vec()).collect(),
+            );
+        }
+        Ok(())
+    }
+
     /// Get the tweezer identifier of the given qubit.
     ///
     /// # Arguments
@@ -621,6 +691,28 @@ impl TweezerDevice {
         None
     }
 
+    /// Returns the two tweezer edges of the device.
+    ///
+    /// And edge between two tweezer is valid only if the
+    /// PhaseShiftedControlledPhase gate can be performed.
+    ///
+    /// # Returns:
+    ///
+    /// * `Vec<(usize, usize)>` - The vector containing the edges.
+    pub fn two_tweezer_edges(&self) -> Vec<(usize, usize)> {
+        let mut edges: Vec<(usize, usize)> = vec![];
+        if let Some(hm) = self
+            .get_current_layout_info()
+            .tweezer_two_qubit_gate_times
+            .get("PhaseShiftedControlledPhase")
+        {
+            for ((start_tw, end_tw), _) in hm.iter() {
+                edges.push((*start_tw, *end_tw));
+            }
+        }
+        edges
+    }
+
     #[inline]
     fn get_current_layout_info(&self) -> &TweezerLayoutInfo {
         self.layout_register
@@ -628,8 +720,14 @@ impl TweezerDevice {
             .expect("Unexpectedly did not find current layout. Bug in roqoqo-qryd.")
     }
 
-    fn is_tweezer_present(&self, tweezer: usize) -> bool {
-        let tweezer_info = self.get_current_layout_info();
+    fn is_tweezer_present(&self, tweezer: usize, layout_name: Option<String>) -> bool {
+        let tweezer_info = if let Some(x) = layout_name {
+            self.layout_register
+                .get(&x)
+                .expect("The specified layout does not exist.")
+        } else {
+            self.get_current_layout_info()
+        };
         let mut present: bool = false;
         for single_qubit_gate_struct in &tweezer_info.tweezer_single_qubit_gate_times {
             if single_qubit_gate_struct.1.contains_key(&tweezer) {
@@ -727,6 +825,39 @@ impl TweezerDevice {
         } else {
             HashMap::new()
         }
+    }
+
+    fn _are_all_shifts_valid(&mut self, pragma: &PragmaShiftQubitsTweezers) -> bool {
+        // Checks for all shifts from pragma:
+        // - if the starting tweezer has any valid shifts associated with it in the device
+        // - if the ending tweezer is contained in the associated valid shifts
+        // - if the device has has a quantum state to move in the starting tweezer position
+        for (shift_start, shift_end) in &pragma.shifts {
+            match self
+                .get_current_layout_info()
+                .allowed_tweezer_shifts
+                .get(shift_start)
+            {
+                Some(allowed_shifts) => {
+                    // Check if each shift is valid and that there's a quantum state to move
+                    if !allowed_shifts
+                        .iter()
+                        .any(|allowed_dest| allowed_dest.contains(shift_end))
+                        && !self
+                            .qubit_to_tweezer
+                            .as_ref()
+                            .unwrap()
+                            .iter()
+                            .any(|(_, twz)| twz == shift_start)
+                    {
+                        return false;
+                    }
+                }
+                // If no shifts are allowed by the device for this tweezer, then it's not valid
+                None => return false,
+            }
+        }
+        true
     }
 }
 
@@ -840,19 +971,19 @@ impl Device for TweezerDevice {
 
     fn change_device(&mut self, hqslang: &str, operation: &[u8]) -> Result<(), RoqoqoBackendError> {
         match hqslang {
-            "PragmaChangeQRydLayout" => {
-                let de_change_layout: Result<PragmaChangeQRydLayout, Box<bincode::ErrorKind>> =
+            "PragmaChangeQRydLayout" => Err(RoqoqoBackendError::GenericError {
+                msg: "Operation not supported in TweezerDevice. Please use PragmaSwitchDeviceLayout.".to_string(),
+            }),
+            "PragmaSwitchDeviceLayout" => {
+                let de_change_layout: Result<PragmaSwitchDeviceLayout, Box<bincode::ErrorKind>> =
                     deserialize(operation);
                 match de_change_layout {
-                    Ok(pragma) => self.switch_layout(&pragma.new_layout().to_string()), // PROBLEM: PragmaChangeQRydLayout indexes by usize, not str
+                    Ok(pragma) => self.switch_layout(pragma.new_layout()),
                     Err(_) => Err(RoqoqoBackendError::GenericError {
-                        msg: "Wrapped operation not supported in ExperimentalDevice".to_string(),
+                        msg: "Wrapped operation not supported in TweezerDevice".to_string(),
                     }),
                 }
-            }
-            "PragmaShiftQRydQubit" => Err(RoqoqoBackendError::GenericError {
-                msg: "Operation not supported in ExperimentalDevice".to_string(),
-            }),
+            },
             "PragmaDeactivateQRydQubit" => {
                 let de_change_layout: Result<PragmaDeactivateQRydQubit, Box<bincode::ErrorKind>> =
                     deserialize(operation);
@@ -862,12 +993,62 @@ impl Device for TweezerDevice {
                         Ok(())
                     }
                     Err(_) => Err(RoqoqoBackendError::GenericError {
-                        msg: "Wrapped operation not supported in ExperimentalDevice".to_string(),
+                        msg: "Wrapped operation not supported in TweezerDevice".to_string(),
+                    }),
+                }
+            },
+            "PragmaShiftQRydQubit" => Err(RoqoqoBackendError::GenericError {
+                msg: "Operation not supported in TweezerDevice. Please use PragmaShiftQubitsTweezers.".to_string(),
+            }),
+            "PragmaShiftQubitsTweezers" => {
+                let de_shift_qubits_tweezers: Result<
+                    PragmaShiftQubitsTweezers,
+                    Box<bincode::ErrorKind>,
+                > = deserialize(operation);
+                match de_shift_qubits_tweezers {
+                    Ok(pragma) => {
+                        // Check if the there are qubits to move
+                        if self.qubit_to_tweezer.is_none() {
+                            return Err(RoqoqoBackendError::GenericError {
+                                msg: "The device qubit -> tweezer mapping is empty: no qubits to shift.".to_string(),
+                            });
+                        }
+                        // Check if the shifts in the operation are valid on the device
+                        if !self._are_all_shifts_valid(&pragma) {
+                            return Err(RoqoqoBackendError::GenericError {
+                                msg: "The PragmaShiftQubitsTweezers operation is not valid on this device."
+                                    .to_string(),
+                            });
+                        }
+                        // Start applying the shifts
+                        if let Some(map) = &mut self.qubit_to_tweezer {
+                            for (shift_start, shift_end) in &pragma.shifts {
+                                if let Some(qubit_to_remove) =
+                                    map.iter()
+                                        .find_map(|(&qbt, &twz)| if twz == *shift_end { Some(qbt) } else { None })
+                                {
+                                    // Remove qubit previously present in target tweezer
+                                    map.remove(&qubit_to_remove);
+                                }
+                                if let Some(qubit_to_move) =
+                                    map.iter()
+                                        .find_map(|(&qbt, &twz)| if twz == *shift_start { Some(qbt) } else { None })
+                                {
+                                    // Move the qubit into the new tweezer
+                                    map.remove(&qubit_to_move);
+                                    map.insert(qubit_to_move, *shift_end);
+                                }
+                            }
+                        }
+                        Ok(())
+                    }
+                    Err(_) => Err(RoqoqoBackendError::GenericError {
+                        msg: "Wrapped operation not supported in TweezerDevice".to_string(),
                     }),
                 }
             }
             _ => Err(RoqoqoBackendError::GenericError {
-                msg: "Wrapped operation not supported in ExperimentalDevice".to_string(),
+                msg: "Wrapped operation not supported in TweezerDevice".to_string(),
             }),
         }
     }